+++ app/src/main/java/uz/behzoddev/restaurantpos/data/local/datasource/StoreDataSource.kt
@@ -6,7 +6,7 @@ import uz.behzoddev.restaurantpos.data.local.models.FoodStore
 
 interface StoreDataSource {
     suspend fun insertStore(foodStore: FoodStore): Long
-    suspend fun updateStore(foodStore: FoodStore): Long
+    suspend fun updateStore(foodStore: FoodStore): Int
     suspend fun deleteStore(foodStore: FoodStore): Int
     fun fetchAllStores(): Flow<List<FoodStore>>
     fun fetchAllStoresById(foodStoreId: Int): Flow<FoodStore>
