+++ app/src/main/java/uz/behzoddev/restaurantpos/domain/interactors/menu/UpdateMenuInteractorImpl.kt
@@ -8,7 +8,7 @@ import javax.inject.Inject
 class UpdateMenuInteractorImpl @Inject constructor(
     @RepositoryBinds private val menuRepository: MenuRepository
 ) : UpdateMenuInteractor {
-    override suspend fun invoke(foodMenu: FoodMenu): Long {
+    override suspend fun invoke(foodMenu: FoodMenu): Int {
         return menuRepository.updateMenu(foodMenu)
     }
 }
