+++ app/src/main/java/uz/behzoddev/restaurantpos/domain/interactors/store/UpdateStoreInteractor.kt
@@ -3,5 +3,5 @@ package uz.behzoddev.restaurantpos.domain.interactors.store
 import uz.behzoddev.restaurantpos.data.local.models.FoodStore
 
 interface UpdateStoreInteractor {
-    suspend operator fun invoke(foodStore: FoodStore): Long
+    suspend operator fun invoke(foodStore: FoodStore): Int
 }
