+++ app/src/main/java/uz/behzoddev/restaurantpos/domain/repositories/ItemRepository.kt
@@ -5,7 +5,7 @@ import uz.behzoddev.restaurantpos.data.local.models.FoodItem
 
 interface ItemRepository {
     suspend fun insertFoodItem(foodItem: FoodItem): Long
-    suspend fun updateFoodItem(foodItem: FoodItem): Long
+    suspend fun updateFoodItem(foodItem: FoodItem): Int
     suspend fun deleteFoodItem(foodItem: FoodItem): Int
     fun fetchAllFoodItems(): Flow<List<FoodItem>>
     fun fetchAllFoodItemsByName(): Flow<List<FoodItem>>
