+++ app/src/main/java/uz/behzoddev/restaurantpos/domain/repositories/StoreRepository.kt
@@ -5,7 +5,7 @@ import uz.behzoddev.restaurantpos.data.local.models.FoodStore
 
 interface StoreRepository {
     suspend fun insertStore(foodStore: FoodStore): Long
-    suspend fun updateStore(foodStore: FoodStore): Long
+    suspend fun updateStore(foodStore: FoodStore): Int
     suspend fun deleteStore(foodStore: FoodStore): Int
     fun fetchAllStores(): Flow<List<FoodStore>>
     fun fetchAllStoresById(foodStoreId: Int): Flow<FoodStore>
